Compare Nested CFA Models

Compares two nested CFA models using chi-square difference test. Also reports AIC and BIC for model comparison.

Usage

compare_cfa_models(fit_small, fit_large, test = "default")

Arguments

fit_small

lavaan fit object for the more constrained (smaller) model

fit_large

lavaan fit object for the less constrained (larger) model

test

Character. Test to use: "default" (chi-square difference), "satorra.bentler", "yuan.bentler.mplus"

Value

A tibble comparing the two models with:

Fit indices for each model

Chi-square difference test results

AIC/BIC comparison

Details

**Nested models:** - Model A is nested in Model B if A is a restricted version of B - Example: 1-factor model is nested in 2-factor model

**Chi-square difference test:** - Tests if the larger model fits significantly better - Deltachi^2 = chi^2_small - chi^2_large - Deltadf = df_small - df_large - If p < .05: Larger model fits significantly better

**AIC/BIC:** - Lower is better - Difference > 10 indicates strong preference

Examples

if (FALSE) { # \dontrun{
# Fit nested models
model_1factor <- "general =~ q1 + q2 + q3 + q4"
model_2factor <- "
  factor1 =~ q1 + q2
  factor2 =~ q3 + q4
"

fit1 <- run_cfa(model_1factor, data = df)
fit2 <- run_cfa(model_2factor, data = df)

# Compare
comparison <- compare_cfa_models(fit1, fit2)
print(comparison)
} # }